+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 Changelog
 =========
 
+Version 0.2
+-----------
+
+  * German locations can have city, zipcode, state and county 
+  * improved search 'algorithm'
+  * new database, new licensing
+  * new `query()` method
+
+
 Version 0.1(dev)
 ----------------
 
-* Introduced change log
+  * Introduced change log